rails api with specs
index endpoints for `groups` and `transactions` resources

Both endpoints are managed by rails controllers under the api namespace.
Specs are written with rspec and fabricator.
Data is serialized with the help of the `blueprinter` gem
A `DateFormatter` service helps format the date acoording to a standart
that is easily parsable with javascript's `Date`.